Update SiteHeader.js
Corrected issue where it was trying to show category list before fetching the category data.

The solution is, when there is no data, then by default: either "loading"  value will be true, or the "error" value will have error. So in both the cases, the component will return valid jsx and will not try to render the category list, which is waiting to be fetched till that time point.